requirements.txt
langgraph 
langchain 
google-generativeai 
fastapi 
fpdf
python-dotenv
uvicorn
langchain-google-genai
